+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.0.5] - 2023-XX-XX
 * Added cdaweb methods that can use cdasws to get the remote file list
+* Maintenance
+  * Added a version cap for numpy (required for cdf interface, revisit before release)
 
 ## [0.0.4] - 2022-11-07
 * Update instrument tests with new test class