+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * ACE MAG
   * ACE SIS
   * ACE SWEPAM
+* New Instruments
+  * DMSP SSUSI EDR-Aurora data
+* Bug Fixes
+  * Updated CDAWeb routines to allow for data stored by year/day-of-year
 * Maintenance
   * Added a version cap for numpy (required for cdf interface, revisit before release)