+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Maintenance
   * Added a version cap for numpy (required for cdf interface, revisit before release)
   * Updated dosctring style for consistency
+  * Removed version cap for xarray
 
 ## [0.0.4] - 2022-11-07
 * Update instrument tests with new test class