+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.0.5] - 2023-XX-XX
+## [0.X.X] - 2023-XX-XX
+* New Instruments
+  * ACE EPAM
+  * ACE MAG
+  * ACE SIS
+  * ACE SWEPAM
 * New Instruments
   * DMSP SSUSI EDR-Aurora data
 * Bug Fixes