+++ CHANGELOG.md
@@ -9,9 +9,17 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * ACE SIS
   * ACE SWEPAM
 * New Instruments
+  * TIMED GUVI
   * DMSP SSUSI EDR-Aurora data
+* Add TIMED GUVI platform to support L1C intensity datasets.
+  * Type of sensor source handled by inst_id with options of
+    spectrograph, imaging
+  * Resolution of dataset handled by tag with
+    low, high
 * Bug Fixes
   * Updated CDAWeb routines to allow for data stored by year/day-of-year
+* Documentation
+  * Added TIMED-GUVI platform
 * Maintenance
   * Added a version cap for numpy (required for cdf interface, revisit before release)
   * Updated actions and templates based on pysatEcosystem docs.