+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.0.5] - 2023-XX-XX
+## [0.X.X] - 2023-XX-XX
 * New Instruments
+  * ACE EPAM
+  * ACE MAG
+  * ACE SIS
+  * ACE SWEPAM
+  * TIMED GUVI L1C intensity data and L2 EDR-Aurora data
   * DMSP SSUSI EDR-Aurora data
-  * TIMED GUVI EDR-Aur data
 * Bug Fixes
   * Updated CDAWeb routines to allow for data stored by year/day-of-year
   * Added missing sub-module imports
@@ -16,6 +20,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Maintenance
   * Added a version cap for numpy (required for cdf interface, revisit before
     release)
+  * Updated actions and templates based on pysatEcosystem docs.
+  * Remove pandas cap on NEP29 tests
+  * Updated dosctring style for consistency
+  * Removed version cap for xarray
+  * Added manual workflow to check that latest RC is installable through test pip
 
 ## [0.0.4] - 2022-11-07
 * Update instrument tests with new test class